refactor: extract chat service revert helpers
Motivation

Start slimming chat_service.py by moving the revert-boundary workflow into its own module without changing the public service API.

Technical Changes

- extract boundary state helpers, undo/redo logic, cleanup of reverted tails, and summary exclusion into app/services/chat_service_revert.py
- keep app/services/chat_service.py as the compatibility surface by importing the shared helpers and wrapping the extracted async functions
- preserve existing test patch points by rebinding the extracted module to the local session_workspace_dir before each wrapper call

Impact

Behavior remains unchanged while reducing one of the central backend hotspot files and creating a clean seam for additional chat service extractions.
